source/docs/004_MCP服务端开发流程/my-mcp-server/pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.4.1",
    "openai>=1.68.0",
    "python-dotenv>=1.0.1",
]
